Use raptor_format_integer() in generating bnode IDs
(raptor_world_default_generate_bnodeid_handler): Use
raptor_format_integer() and tidy code.

raptor_world field default_generate_bnodeid_handler_prefix_length
is an integer; long enough.

+++ src/raptor_internal.h
@@ -1417,7 +1417,7 @@ struct raptor_world_s {
 
   int default_generate_bnodeid_handler_base;
   char *default_generate_bnodeid_handler_prefix;
-  size_t default_generate_bnodeid_handler_prefix_length;
+  unsigned int default_generate_bnodeid_handler_prefix_length;
 };
 
 /* raptor_www.c */
